Added enemy AI/waypoints to minigame scene
The minima scene now incorporated a knavish on the floor, and ‘smart’
waypoints that procedurally generate the next waypoint for the enemy
AI.  Each waypoint (transparent and formless, but registers contact
with the enemy) has an array of possible destinations from that point.
The waypoint object randomly selects the next waypoint and sends this
information to the enemy AI via the SendMessage function.  In the final
maze, each ‘maze square’ will have to have one of these waypoint
objects embedded in it.